pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-eval"
version = "0.1.0"
description = "Codebase RAG evaluation harness"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Code RAG Eval" }]
dependencies = [
  "typer>=0.12",
  "PyYAML>=6.0",
  "gitpython>=3.1",
  "rich>=13.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.6"
]

[project.scripts]
rag-eval = "rag_eval.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

